eatgo-api/src/main/java/kr/co/project/eat/application/RestaurantService.java
package kr.co.project.eat.application;

import kr.co.project.eat.domain.Restaurant;
import kr.co.project.eat.domain.RestaurantRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class RestaurantService {

    @Autowired
    RestaurantRepository restaurantRepository;

    public RestaurantService(RestaurantRepository restaurantRepository) {
        this.restaurantRepository = restaurantRepository;
    }

    public Restaurant getRestaurant(Long id){

        Restaurant restaurant =restaurantRepository.findById(id);
        return restaurant;
    }

}
